Fixes bug where DailyPerformance#last_30_days is run early in year
I chose to work with integers to represent days early on when modeling
the data.  I don't regret that choice, but forgot to think about future/
  past years.  Now some minor calculations are run when quering the
data; if we're in said corner case, we query for all the days of this
year, and the last applicable days of the prior year.